Bug fix: the Gtk+ Provider needs to use forkOS, not forkIO.
When using forkIO, a Worker thread could be created and begin looping
over the task, but a loop would never until some event occurred. So
clicking the mouse would cause the Worker loop to be evaluated, but it
would never loop without some other event handler thread evaluating
first. To solve this problem it was necessary to use 'forkOS' and
build Happlets with the -threaded option set in GHC-OPTIONS.