Initial implementation of Happlets.Audio API.
This API has been informed by a lot of work I have done in parallel
for a related project (the "procgen" package), some of the API names
were taken directly from that project. So I am already fairly
confident that this API is already pretty close to it's final form.