I have decided not to expose the underlying pixel buffer data type
The purpose of Happlets is to provide a canvas that responds to
events. The canvas data type and all of it's event handlers are
encapsulted in whatever "provider" data type is used to instantiate
the various provider typeclasses. To provide abstractions to the
underlying pixel buffer would be like having a different "provider" to
which only some of the happlet typeclasses applied. This doesn't make
much sense to me. If you need only drawing operatinos and nothing
else (no event handling), then why use Happlets at all?