Redefine liftGUIProvider, ProviderStateLock, runProviderOnLock
Now, the 'liftGUIProvider', 'runProviderOnLock', and 'providerLiftGUI'
functions all are more orthogonal in their types and functionality, it
is clear from the types when to use which function. The
'providerRunSync' function is now by default a synonym for
'runProviderOnLock' unless it is overridden, but the comments make
clear that if 'providerRunSync' is to be overridden
'runProviderOnLock' must still be called.