attachProvider renamed back to attachWindow, args ordering changed
This function was originally called 'attachWindow' and I probably did
a automated find/replace to change the name of the function to
'attachProvider'. But I think maybe 'attachWindow' is probably a more
intuitive name.

The arguments were rearranged to take the provider lock first and a
boolean indicating whether to make the window immeiately visible
second -- prior to this commit those first to arguments were
flipped. IO actions typically take the environment value first, before
this commit a boolean was the first argument. This was unconventional,
so I changed it.

Also, fixed up comments referring to 'windowChangeHapplet', that
function was renamed to 'changeRootHapplet'.