Rename getVal/setVal to getEnv/setEnv
Although it may seem counter intuitive when defining function of type
'Variable' that the getter and setter would have the term "env" in the
name (aren't you setting the value of the variable, not an
environment?), when, however, you actually use these functions, the
term 'getEnv' seems to be a bit more intuitive to me. When you read
the code and you see 'getEnv' or 'setEnv', it is immediately obvious
to anyone with shell scripting experience what these functions are
doing.