Made a MUCH needed change to the forkGUI API, adding SendGUISignal
Now the caller of 'forkGUI' has a choice of when and how often to
evaluate a 'GUI' function. This means parallel threads of computation
effectively become agents that programmatically send signals to the
GUI. The signals themselves have no data structure or any
pattern-matchable form, they are simply 'GUI' functions that are
evaluated, but this is enough to create threads that send signals to
the GUI when (for example) receiving messages on a file descriptor.