Merged Config.hs and Provider.hs into Initialize.hs
Each of these modules were small and inter-related, it is impractical
to split them up into different modules.

Also renamed 'Variable' to 'StateConfig', renamed 'setEnv' to
'setConfig', and renamed 'getEnv' to 'getConfig'. I did this because I
realized the 'getEnv' and 'setEnv' names might conflict with the
POSIX-related functions for managing environment variables.