Re-define Draw2DPrimitive, remove Draw2DShape
The Draw2DPrimitive included a Draw2DShape and instructions for stroke
and fill. However not all Draw2DShapes were fillable, namely the
Line2D primitive and the Draw2DReset primitive. So I removed the
Draw2DShape data type and created the Draw2DFillStroke type, and now
each primitive is associated with it's own PaintSource or
Draw2DFillStroke functino.