SampCoord is no longer a newtype in it's own module
This is because the Vector library is indeed quite finicky. The
unboxed types are a walled garden, and trying to introduce a unboxed
newtype is unreasonably difficult, at least as far as I was able to
figure out by reading the source code (maybe there is an easy way that
I don't know about, I couldn't figure it out, though).

Rather than swtich to Repa or something with a better designed, more
modern type system, for the time being I will just remove the newtype
and use the unboxed type directly.