Delete special animation event handler 'AnimateAction'
All event handlers now take the form of 'EventAction'. There is no
need for an event handler specialized to animation events. Before I
thought it would have been convenient if the Role of an actor had
keept track of the animation start time or elapsed time. Now it is
clear that this functionality should NOT be included in this module
any more than the functionality for drawing to the canvas or tracking
whether a click occurs within the visible region of an
Actor. Functionality like this shall be defined by the users of the
API for each 'Actor', and the visibility and the animation time can be
held within the @model@ of the 'Actor'.