Rename "mouse events" to "mouse signals"
I realized that the "Happlets.Control.Mouse" module establishes a
low-level signal handler, not an event handler, as it would be
described in a FRP framework. The 'Mouse' data type contained
information about whether a signal was mouse-down or mouse-up, and
thus is a signal, not an event. This can be tranlated to events using
an FRP framework. In the case of the "Happlets.Scene" module, an FSA
is defined to translate these signals to events and trigger event
handler 'Script' functions.