Overhauled of Scene event handlers
This comes after the realization that an app should be constructed
from heirarchy or network of 'TypedActors' are, and the untyped Actors
are separate objects that delegate events to TypedActors, and that
Scenes are collections of these delegates that can broadcast events to
all delegates. Now, propagation of events should be much more correct
than it was before.