Complete overhaul of Actors and Scenes
I realized that typed 'Actor's can contain other typed 'Actor's
elements within them, and can delegate events to these element
'Actor's, meaning there is really no need for a 'Scene' at all. But,
the 'Scene' type containing many untyped 'Actor's and delegating
events to all actors within mechanically might also be useful, but it
should be used as a typed 'Actor' that contains the 'Scene' within.

So I renamed the untyped 'Actor' to 'Presence', and I renamed
'TypedActor' to simply 'Actor' since it is now the main type to be
used in an 'Act'. 'Scene' is still 'Scene' but it is just another
'Actor' data type, it is not a required built-in component of a
'Script' evaluation context anymore.